+++ BTCPayServer.Plugins.LNbank/Services/Wallets/BoltCardsQuery.cs
@@ -7,7 +7,7 @@ public class BoltCardsQuery
     public string BoltCardId { get; set; }
     public string WithdrawConfigId { get; set; }
     public bool IncludeWithdrawConfig { get; set; }
-    public bool IncludeTransactions { get; set; }
+    public bool IncludeWallet { get; set; }
     public bool IncludeSoftDeleted { get; set; }
     public bool IsServerAdmin { get; set; }
     public BoltCardStatus? Status { get; set; }